Literally the working build after maybe 24 hours of work
Works pretty much flawlessly, but a couple caveats (that will be fixed):
The Song Name is currently based on the Song Shown in Clone Hero, which would not be the same in FNF.
Does NOT Generate a modfolder, needs to be manually moved into a mod that you setup for it.
2 Extra bugs, one where the first section in the song gets filled with garbage notes.  No Idea whats going on there
And There was this weird one off bug where the sustain of like 5 notes got extended to the position value of the note that came after it?  No idea wtf happened there.  Hopefully it won't.  WIll update later with fixes for all of this